Move CSV translators to separate class
